Refactor so de-integration can be used elsewhere in CocoaPods
