+++ lib/cocoapods/deintegrate/gem_version.rb
@@ -1,0 +1,3 @@
+module CocoapodsDeintegrate
+  VERSION = '0.2.1'
+end
