+++ lib/cocoapods_plugin.rb
@@ -1,2 +1,4 @@
-require 'pod/command/deintegrate'
-
+module Pod
+  require 'cocoapods_deintegrate'
+  require 'cocoapods/command/deintegrate'
+end
