+++ spec/spec_helper.rb
@@ -26,4 +26,3 @@ module Pod
     end
   end
 end
-
